+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Adds related articles [v1.0.1]
+- Adds support for product area labels
+- Adds support for related articles
+
 ## First Commit [v 1.0.0]
 - Setup of the project
 - Basic modules added to parse Prismic response